fix: modified name file inverse to ioc and refactor types
