+++ package.json
@@ -37,6 +37,7 @@
     "ts-jest": "29.1.1",
     "ts-node": "10.0.0",
     "tsc-alias": "1.8.8",
+    "tsconfig-paths": "^4.2.0",
     "typescript": "4.9.5"
   },
   "dependencies": {
