+++ src/api/types.ts
@@ -1,0 +1,4 @@
+export const TYPES_API = {
+  googleMaps: Symbol('GoogleMaps'),
+  postmon: Symbol('Postmon'),
+};
