+++ src/index.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata';
 
-import { configureContainer } from '@configs/inverse';
+import { configureContainer } from '@configs/ioc';
 configureContainer();
 
 export { cityService } from '@services/index';
