+++ src/services/CityService/types.ts
@@ -1,0 +1,3 @@
+export const TYPES_SERVICE = {
+  cityService: Symbol('CityService'),
+};
